+++ src/light.cc
@@ -276,17 +276,17 @@ void restore_light_sources(int fd) {
 /* Relink all lights that are so marked. */
 void relink_light_sources(boolean ghostly) {
     char which;
-    unsigned nid;
+    unsigned long nid;
     light_source *ls;
 
     for (ls = light_base; ls; ls = ls->next) {
 	if (ls->flags & LSF_NEEDS_FIXUP) {
 	    if (ls->type == LS_OBJECT || ls->type == LS_MONSTER) {
 		if (ghostly) {
-		    if (!lookup_id_mapping((unsigned)ls->id, &nid))
+		    if (!lookup_id_mapping((unsigned long)ls->id, &nid))
 			impossible("relink_light_sources: no id mapping");
 		} else
-		    nid = (unsigned) ls->id;
+		    nid = (unsigned long) ls->id;
 		if (ls->type == LS_OBJECT) {
 		    which = 'o';
 		    ls->id = (genericptr_t) find_oid(nid);
@@ -295,7 +295,7 @@ void relink_light_sources(boolean ghostly) {
 		    ls->id = (genericptr_t) find_mid(nid, FM_EVERYWHERE);
 		}
 		if (!ls->id)
-		    impossible("relink_light_sources: cant find %c_id %d",
+		    impossible("relink_light_sources: cant find %c_id %ld",
 			       which, nid);
 	    } else
 		impossible("relink_light_sources: bad type (%d)", ls->type);
